Add input validation and optimized committed
Add statements in click events that validates inputs of zeros, dots and math operators.
Removed variables assigned to unused dom objects
Removed variables which linked input with display dom objects which allowed to remove unnacesary code and simplified data validation.
Replaced 00 with 0 in output div index html to simplify data validation. Now both screens starts with 0.
This needs to be further adjusted with css to highlight difference between screens for users.
backspace function is not going to be used twice. Therefore, it is moved into backspace click event to reduce the code.